Fix use of ovs-wrapper
The `ovs-ctl` script will do automatic initialization and
preservation of the System ID, only when the `--system-id=random`
argument is passed.

It is not allowed to use the equal (`=`) sign in a snap app
command, and environment variables are not passed into the
generated systemd services files.

So our only option to allow this to work is unfortunately by using
a wrapper script to start the OVS daemons.

Previously this was only used for the `ovsdb-server` service,
probably because it starts first, but this causes problem when
the snap is restarted.